requirements.txt
grpcio==1.76.0
grpcio-tools==1.76.0
protobuf==6.33.1
psycopg2==2.9.11
pyodbc==5.3.0
typing-extensions==4.15.0
pandas
